# Implement one-batch lookahead for index enumerators