@Swiddis 2025-10-07
Turns out I flipped the benchmark in my head, so this is overall a regression -- going to put back in draft and figure out a better approach